% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manualComplexFeatureAnnotation.R
\name{annotateComplexFeatures}
\alias{annotateComplexFeatures}
\title{Annotate complex features using the manually annotated CORUM
complexes.}
\usage{
annotateComplexFeatures(detected.features, manual.annotations,
  with.sec = TRUE, feature.vicinity.tol = 5)
}
\arguments{
\item{detected.features}{A data.table holding the features as output
by cprophet PC.}

\item{manual.annotations}{A data.table with the columns 'complex_id', 'rt'.
The built in data sets like `manual.annotation.full` can be used
here.}

\item{feature.vicinity.tol}{A numeric value that specifies how near an
experimental feature has to be to a manually
annotated one to count as a true positive}
}
\value{
The same data.table with a new column 'is_true_positive'.
}
\description{
Each row will get an additional column `is_true_positive`.
}

