% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manualComplexFeatureAnnotation.R
\name{annotateComplexFeaturesAlt}
\alias{annotateComplexFeaturesAlt}
\title{Annotate complex features by simply checking if there is a manual annotation
nearby. This function is faster than `annotateComplexFeatures`, which loops
through all the annotations and ensures that only one detected feature can
be assigned to one manually annotated feature.}
\usage{
annotateComplexFeaturesAlt(detected.features, manual.annotations,
  feature.vicinity.tol = 5)
}
\description{
Annotate complex features by simply checking if there is a manual annotation
nearby. This function is faster than `annotateComplexFeatures`, which loops
through all the annotations and ensures that only one detected feature can
be assigned to one manually annotated feature.
}

