% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manualComplexFeatureAnnotation.R
\name{apexStringToDF}
\alias{apexStringToDF}
\title{Split a list of apexes like '1,2,3,4' and create a data.frame
data.frame(rt=c(1, 2, 3, 4), complex.id=X, apex.type=Y).
If no split was possible, NULL is returned.}
\usage{
apexStringToDF(complex.id, sep.apexes, apex.type)
}
\arguments{
\item{complex.id}{A string identifying the complex.}

\item{sep.apexes}{A string of comma-separated numbers.}

\item{apex.type}{A string identifying the type of apex.}
}
\value{
Either NULL or a data.frame.
}
\description{
Split a list of apexes like '1,2,3,4' and create a data.frame
data.frame(rt=c(1, 2, 3, 4), complex.id=X, apex.type=Y).
If no split was possible, NULL is returned.
}

