% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manualComplexFeatureAnnotation.R
\name{assessComplexFeatures}
\alias{assessComplexFeatures}
\title{Check for each feature in `detected.features` if there is another feature in
`true.features` that is close to it. The endresult is a DF of feature
retention times that are flagged with either 'FP', 'TP', or 'FN'.}
\usage{
assessComplexFeatures(true.features, detected.features,
  feature.vicinity.tol = 5)
}
\arguments{
\item{true.features}{data.table of true, manually annotated features.
The DF must have the columns: 'complex_id', 'rt'.}

\item{detected.features}{data.table of detected features.
The DF must have the columns: 'complex_id', 'rt'.}

\item{feature.vicinity.tol}{A number that indicates how close an
experimentally determined feature has to be to a manually annotated one, to
still count as a true positive.}
}
\value{
A data.table with the columns: 'complex_id', 'rt', 'type', where
type is of type character an is either 'FP', 'FN', 'TP'.
}
\description{
Check for each feature in `detected.features` if there is another feature in
`true.features` that is close to it. The endresult is a DF of feature
retention times that are flagged with either 'FP', 'TP', or 'FN'.
}

