% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manualComplexFeatureAnnotation.R
\name{calcFPR}
\alias{calcFPR}
\title{Calculate the false positive rate (== 1 - specificity)}
\usage{
calcFPR(dt)
}
\description{
Calculate the false positive rate (== 1 - specificity)
}

