% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benchmarkAnalysisCC.R
\name{calculateDetectionStatsCC}
\alias{calculateDetectionStatsCC}
\title{Calculate stats for the CC workflow.}
\usage{
calculateDetectionStatsCC(gs.run.directory, manual.annotations,
  sec.range = 3:83, input.complexes = NULL, d.score.cutoffs = c(-2, -1, 0,
  1, 2, 3), window.range = c(2, 4, 8, 12, 18))
}
\description{
Calculate stats for the CC workflow.
}

