% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benchmarkAnalysisPC.R
\name{calculateDetectionStatsPC}
\alias{calculateDetectionStatsPC}
\title{Calculate stats for the PC workflow.}
\usage{
calculateDetectionStatsPC(gs.run.directory, manual.annotations,
  sec.range = 3:83, window.range = c(2, 4, 8, 12, 18))
}
\description{
Calculate stats for the PC workflow.
}

