% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benchmarkAnalysisPC.R
\name{computeSiteLocalizationRate}
\alias{computeSiteLocalizationRate}
\title{Compute for a data table of detected features how many of those that belong
to a true complex also have a annotated feature that is in the vicinity of
them.}
\usage{
computeSiteLocalizationRate(detected.features, manual.annotations,
  feature.vicinity.tol = 5)
}
\description{
Compute for a data table of detected features how many of those that belong
to a true complex also have a annotated feature that is in the vicinity of
them.
}

