% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slidingWindowCorrelation.R
\name{computeWindowedCorrelation}
\alias{computeWindowedCorrelation}
\title{Compute the correlation between rows in a matrix considering only a window.
This is a helper method.}
\usage{
computeWindowedCorrelation(trace.mat, start.window.idx, window.size)
}
\arguments{
\item{trace.mat}{A numeric matrix where rows correspond to the different
traces.}

\item{start.window.idx}{Column index where the window starts.}

\item{window.size}{Size of the window. Numeric.}
}
\value{
The mean correlation between all traces within this window.
}
\description{
Compute the correlation between rows in a matrix considering only a window.
This is a helper method.
}

