% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/complexCentricWorkflow.R
\name{convertToPCComplexFeatureFormat}
\alias{convertToPCComplexFeatureFormat}
\title{Convert the output CC cprophet to a format that looks like it was output by
the PC workflow.}
\usage{
convertToPCComplexFeatureFormat(complex.features, corum.complex.protein.assoc)
}
\description{
Convert the output CC cprophet to a format that looks like it was output by
the PC workflow.
}

