% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataConversion.R
\name{convertToPeptideCCData}
\alias{convertToPeptideCCData}
\title{Convert a peptide trace matrix as is input to cprophet to the same matrix
that associates peptides to complexes, rather than to proteins.
The result of this transformation can then be used as the input for a CC
workflow that tries to create complex features directly from peptides.}
\usage{
convertToPeptideCCData(pc.input)
}
\arguments{
\item{pc.input}{The data.table that is normally input to cprophet. It has
columns: \itemize{
 \item peptide_id
 \item protein_id
 \item The remaining columns run across the SEC dimension.
}}
}
\description{
Convert a peptide trace matrix as is input to cprophet to the same matrix
that associates peptides to complexes, rather than to proteins.
The result of this transformation can then be used as the input for a CC
workflow that tries to create complex features directly from peptides.
}

