% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corum.complex.protein.assoc}
\alias{corum.complex.protein.assoc}
\title{A data.table of associations between corum complexes and their
protein subunits. Complexes that have multiple possible subunit
compositions are split into separate complexes that are numbered
with `-1`, `-2`, etc. suffixes.}
\format{\preformatted{Classes ‘data.table’ and 'data.frame':	9701 obs. of  3 variables:
 $ complex_id  : chr  "230" "5450" "867" "5274" ...
 $ complex_name: chr  "Mediator complex" "Mediator complex" "TRAP-SMCC mediator complex" "Cell-cell junction complex (ARHGAP10-CTNNA1)" ...
 $ protein_id  : chr  "A0JLT2" "A0JLT2" "A0JLT2" "A1A4S6" ...
 - attr(*, ".internal.selfref")=<externalptr> 
 - attr(*, "sorted")= chr "protein_id"
}}
\usage{
corum.complex.protein.assoc
}
\description{
A data.table of associations between corum complexes and their
protein subunits. Complexes that have multiple possible subunit
compositions are split into separate complexes that are numbered
with `-1`, `-2`, etc. suffixes.
}
\keyword{datasets}

