% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manualComplexFeatureAnnotation.R
\name{createApexDF}
\alias{createApexDF}
\title{Given a DF with columns 'complex_id', and another column holding
comma-separated strings of numbers, create a long list style DF
where each row corresponds to a number.}
\usage{
createApexDF(annotations, apex.col.name)
}
\description{
Given a DF with columns 'complex_id', and another column holding
comma-separated strings of numbers, create a long list style DF
where each row corresponds to a number.
}

