% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/complexCentricWorkflow.R
\name{createCCInputData}
\alias{createCCInputData}
\title{Create input data wor the CC workflow.
All this function essentially does is to take a long list peptide trace
data.table, produce the protein traces, annotate them with their complex
membership, rename all the columns such that they are compatible with
cprophet.}
\usage{
createCCInputData(peptide.traces.long, out.file.name, complex.protein.assoc)
}
\description{
Create input data wor the CC workflow.
All this function essentially does is to take a long list peptide trace
data.table, produce the protein traces, annotate them with their complex
membership, rename all the columns such that they are compatible with
cprophet.
}

