% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manualComplexFeatureAnnotation.R
\name{createManualComplexAnnotations}
\alias{createManualComplexAnnotations}
\title{Given a DF with columns 'complex_id', and another column holding
comma-separated strings of numbers, create a long list style DF
where each row corresponds to a number.}
\usage{
createManualComplexAnnotations(annotations, apex.col.name)
}
\arguments{
\item{annotations}{A data.frame with columns 'complex_id' and a second
column with comma-separated retention times of the
complex features.}

\item{apex.col.name}{The name of the column that holds the retention
times.}
}
\value{
A data.table with the following columns: 'complex_id', 'rt'.
}
\description{
Given a DF with columns 'complex_id', and another column holding
comma-separated strings of numbers, create a long list style DF
where each row corresponds to a number.
}
\examples{
manual.annotations.raw <- readManualAnnotationFile('somefile.tsv')
manual.annotations <-
    createManualComplexAnnotations(manual.annotations.raw$annotations,
                                   'apexes_partially_observed')
}

