% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manualComplexFeatureAnnotation.R
\name{createMergedList}
\alias{createMergedList}
\title{Merge the RTs for apexes of `apex.type` for two DTs.
dt1 is treated as the reference DT.}
\usage{
createMergedList(dt1, dt2, apex.type)
}
\description{
Merge the RTs for apexes of `apex.type` for two DTs.
dt1 is treated as the reference DT.
}

