% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deconvComplexDetection.R
\name{deconvProteinTraces}
\alias{deconvProteinTraces}
\title{Deconvolute a wide format data.table of protein traces using Gaussian
deconvolution.}
\usage{
deconvProteinTraces(protein.traces)
}
\description{
Deconvolute a wide format data.table of protein traces using Gaussian
deconvolution.
}

