% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slidingWindowSubgroupDetection.R
\name{detectGroupsWithinWindow}
\alias{detectGroupsWithinWindow}
\title{Detect subgroups within a window. This method should be called by
`slidingWindow`.}
\usage{
detectGroupsWithinWindow(tracemat, corr.cutoff, with.plot = F)
}
\value{
A vector of the same length as the number of proteins. Each element
        corresponds to a integer-based cluster label.
}
\description{
Detect subgroups within a window. This method should be called by
`slidingWindow`.
}

