% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slidingWindowSubgroupDetection.R
\name{detectSubgroupsSW}
\alias{detectSubgroupsSW}
\title{Detect subgroups of proteins within a matrix of protein intensity traces by
sliding a window across the SEC dimension. Within each window, the traces
that correlate well are clustered together.}
\usage{
detectSubgroupsSW(trace.mat, protein.names, corr.cutoff = 0.75,
  window.size = 15, with.plot = F)
}
\arguments{
\item{trace.mat}{A numeric matrix where rows correspond to the different
traces.}

\item{window.size}{Size of the window. Numeric.}

\item{start.window.idx}{Column index where the window starts.}

\item{score}{The type of correlation to use. A string with value 'pearson'
or 'diff'.}
}
\value{
The mean correlation between all traces within this window.
}
\description{
Detect subgroups of proteins within a matrix of protein intensity traces by
sliding a window across the SEC dimension. Within each window, the traces
that correlate well are clustered together.
}

