% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{e4.input.proteins}
\alias{e4.input.proteins}
\title{A list of all proteins that are found in the traces table
e4.peptide.traces.wide. Each protein has a protein_mw that is taken from
UniProt and a protein_concentration that is either imputed or computed by
aLFQ.}
\format{\preformatted{Classes ‘data.table’ and 'data.frame':	7304 obs. of  3 variables:
 $ protein_id           : chr  "A0AVT1" "A0FGR8" "A0MZ66" "A1L0T0" ...
 $ protein_mw           : num  117970 102357 71640 67868 268571 ...
 $ protein_concentration: num  59316 84070 50108 63234 45352 ...
 - attr(*, ".internal.selfref")=<externalptr> 
}}
\usage{
e4.input.proteins
}
\description{
A list of all proteins that are found in the traces table
e4.peptide.traces.wide. Each protein has a protein_mw that is taken from
UniProt and a protein_concentration that is either imputed or computed by
aLFQ.
}
\keyword{datasets}

