% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{e4.peptide.traces.wide}
\alias{e4.peptide.traces.wide}
\title{A wide format data.table of peptide trace intensities.}
\format{A data.table with 83266 rows and the following columns:
\itemize {
 \item protein_id: the id of the protein to which this peptide might belong
 \item peptide_id: the id of the peptide itself
 \item 3: intensity measurements for sec fraction 3
 \item 4: intensity measurements for sec fraction 4
 \item ...: further measurements
 \item 83: intensity measurements for sec fraction 84
}}
\usage{
e4.peptide.traces.wide

e4.peptide.traces.wide
}
\description{
The settings applied to the raw data are:
\itemize {
 \item m_score < 0.01
 \item no requant values
 \item no decoy values
}

The settings applied to the raw data are:
\itemize {
 \item m_score < 0.01
 \item no requant values
 \item no decoy values
}
}
\keyword{datasets}

