% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imputePartialPeakgroupCompleteness.R
\name{imputePartialPeakgroupCompletness}
\alias{imputePartialPeakgroupCompletness}
\title{Estimate the completness of a peak group that was annotated as being
partially complete.}
\usage{
imputePartialPeakgroupCompletness(manual.annotations.partial, protein.traces,
  window.size = 14, corr.cutoff = 0.7)
}
\arguments{
\item{manual.annotations.partial}{A manual annotation data.table}

\item{protein.traces}{Wide format protein traces with columns:
\itemize{
  \item complex_id
  \item protein_id
  \item complex_name
  \item additional columns for each sec rt
}}

\item{window.size}{The window in SEC that is considered around each manually
annotated RT.}

\item{corr.cutoff}{Above which value a group of traces is considered to be well
correlated.}
}
\description{
Estimate the completness of a peak group that was annotated as being
partially complete.
}
\examples{
peptraces <- widePepTracesToLong(e4.peptide.traces.wide.filtered)
prottraces <- produceProteinTraces(peptraces)
prottraces.wc <- annotateProteinTraces(prottraces, corum.complex.protein.assoc)
imputePartialPeakgroupCompletness(manual.annotations.partial, prottraces.wc)
}

