% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataImputation.R
\name{imputeProteinConcentrations}
\alias{imputeProteinConcentrations}
\title{Impute concentrations for proteins based on their total intensity across the
SEC dimension and a list of proteins for which both SEC intensity
information and concentration estimates exist.}
\usage{
imputeProteinConcentrations(peptide.traces, protein.conc.estimates,
  protein.ids.file)
}
\arguments{
\item{peptide.traces}{A long list format data.table of peptide traces.}

\item{protein.conc.estimates}{A data.table holding the existing protein
concentrations. It has the format:
\itemize{
  \item protein_id
  \item protein_mw
  \item protein_concentration
}}

\item{protein.ids.file}{A file path where the protein ids are stored for
which a manual uniprot lookup has to be made.}
}
\value{
A data.table holding all proteins with their molecular weight and
        concentration.
}
\description{
Furthermore, missing protein molecular weights are also added.
}
\examples{
peptide.traces <- widePepTracesToLong(e4.peptide.traces.wide)
protein.ids.file <- '~/Desktop/proteinmw.txt'
imputeProteinConcentrations(peptide.traces, HEK293.protein.conc.estimates,
                            protein.ids.file)
}

