% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataConversion.R
\name{longPepTracesToWide}
\alias{longPepTracesToWide}
\title{Convert a long list peptide trace to a wide format.
export}
\usage{
longPepTracesToWide(dt)
}
\description{
Convert a long list peptide trace to a wide format.
export
}

