% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataInput.R
\name{longProtTracesToWide}
\alias{longProtTracesToWide}
\title{Convert a long list protein trace to a wide format.
export}
\usage{
longProtTracesToWide(dt)
}
\description{
Convert a long list protein trace to a wide format.
export
}

