% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manualComplexFeatureAnnotation.R
\name{makeROCWithSEC}
\alias{makeROCWithSEC}
\title{Make ROC curves and treat every SEC position as a possible feature.}
\usage{
makeROCWithSEC(detected.features, true.positive.features, cutoffs,
  feature.vicinity.tol = 5)
}
\arguments{
\item{detected.features}{A dataframe with the columns: 'complex_id', 'rt'}

\item{true.positive.features}{A dataframe with the columns: 'complex_id', 'rt'}

\item{cutoffs}{A numeric vector of apex_mw_fit values that should be used to
define if a feature passes the molecular weight test or not}

\item{feature.vicinity.tol}{A numeric value that specifies how near an
experimental feature has to be to a manually
annotated one to count as a true positive}
}
\description{
Make ROC curves and treat every SEC position as a possible feature.
}

