% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlsGaussianSum.R
\name{makeStartParams}
\alias{makeStartParams}
\title{Create a named vector of starting parameters for an nlsGaussiamSum
model with `n.components` gaussians.
For n.components == 2, the value of this function could look like this:
c(a1=40000, s1=4, mu1=30, a2=40000, s2=4, mu2=60)
The exact values depend on the maximum intensity `max.y` that was observed
in a trace and the maximum value on the horizontal axis.
The parameters result in `n.components` gaussian curves that are spread
evenly across the range of x-values.}
\usage{
makeStartParams(n.components, max.x, max.y)
}
\arguments{
\item{n.components}{The number of gaussian curves to use.}

\item{max.x}{The maximum value in the x-range.}

\item{max.y}{The maximum value in the y-range.}
}
\value{
A named vector of parameters that can be plugged into the fitting
function used internally.
}
\description{
Create a named vector of starting parameters for an nlsGaussiamSum
model with `n.components` gaussians.
For n.components == 2, the value of this function could look like this:
c(a1=40000, s1=4, mu1=30, a2=40000, s2=4, mu2=60)
The exact values depend on the maximum intensity `max.y` that was observed
in a trace and the maximum value on the horizontal axis.
The parameters result in `n.components` gaussian curves that are spread
evenly across the range of x-values.
}

