% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manual.annotations.full.complete}
\alias{manual.annotations.full.complete}
\title{The merged manual CORUM annotations with only those features that consist of
a peak group that was fully observed.
In addition this dataset holds only those fully observed peak groups that
are also complete in the sense that all protein subunits as annotated in
CORUM are actually part of the peak group.}
\format{\preformatted{Classes ‘data.table’ and 'data.frame':	487 obs. of  2 variables:
 $ complex_id: Factor w/ 591 levels "635","1169","730",..: 1 1 2 2 2 5 10 11 22 25 ...
 $ rt        : num  18 20 12 21 17 18 40 40 35 10 ...
 - attr(*, ".internal.selfref")=<externalptr> 
}}
\usage{
manual.annotations.full.complete
}
\description{
The merged manual CORUM annotations with only those features that consist of
a peak group that was fully observed.
In addition this dataset holds only those fully observed peak groups that
are also complete in the sense that all protein subunits as annotated in
CORUM are actually part of the peak group.
}
\keyword{datasets}

