% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manual.annotations.full}
\alias{manual.annotations.full}
\title{The merged manual CORUM annotations with only those features that consist of
a peak group that was fully observed.}
\format{\preformatted{Classes ‘data.table’ and 'data.frame':	1090 obs. of  2 variables:
 $ complex_id: Factor w/ 591 levels "635","1169","730",..: 1 1 2 2 2 3 3 4 4 5 ...
 $ rt        : num  18 20 12 21 17 11 18 11 18 18 ...
 - attr(*, ".internal.selfref")=<externalptr> 
 - attr(*, "index")= atomic  
  ..- attr(*, "__complex_id")= int 
}}
\usage{
manual.annotations.full
}
\description{
The merged manual CORUM annotations with only those features that consist of
a peak group that was fully observed.
}
\keyword{datasets}

