% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manual.annotations.raw.mheusel}
\alias{manual.annotations.raw.mheusel}
\title{The raw manual complex annotations by mheusel
These annotations are based on the traces
summed up protein traces from
e4.peptide.traces.wide.filtered.
A already processed and merged  version exists in the data frame
\itemize {
 \item manual.annotations.partial
 \item manual.annotations.full
 \item manual.annotations.full.complete
}}
\format{\preformatted{ chr "~/Dev/MAScripts/manual_complex_annotation/151109_SEC-SWATH_CORUM_manual_annotation_mhe.tsv"
}}
\usage{
manual.annotations.raw.mheusel
}
\description{
The raw manual complex annotations by mheusel
These annotations are based on the traces
summed up protein traces from
e4.peptide.traces.wide.filtered.
A already processed and merged  version exists in the data frame
\itemize {
 \item manual.annotations.partial
 \item manual.annotations.full
 \item manual.annotations.full.complete
}
}
\keyword{datasets}

