% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manualComplexFeatureAnnotation.R
\name{mergeManualComplexAnnotations}
\alias{mergeManualComplexAnnotations}
\title{Merge the RTs for apexes of `apex.type` for two DTs.
dt1 is treated as the reference DT.
Merge two data.tables where each row corresponds to manually annotated
complex feature.
The table must have the columns: 'complex_id' and 'rt'.}
\usage{
mergeManualComplexAnnotations(dt1, dt2)
}
\arguments{
\item{dt1}{The first DT.}

\item{dt2}{The second DT.}
}
\description{
Merge the RTs for apexes of `apex.type` for two DTs.
dt1 is treated as the reference DT.
Merge two data.tables where each row corresponds to manually annotated
complex feature.
The table must have the columns: 'complex_id' and 'rt'.
}
\examples{
manual.annotations.1.raw <- readManualAnnotationFile(annotations.1.raw)
manual.annotations.2.raw <- readManualAnnotationFile(annotations.2.raw)
manual.annotations.1 <-
    createManualComplexAnnotations(manual.annotations.1.raw$annotations,
                                   'apexes_partially_observed')
manual.annotations.2 <-
    createManualComplexAnnotations(manual.annotations.2.raw$annotations,
                                   'apexes_partially_observed')
manual.annotations <- mergeManualComplexAnnotations(manual.annotations.1,
                                                    manual.annotations.2)
}

