% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manualComplexFeatureAnnotation.R
\name{mergeRTs}
\alias{mergeRTs}
\title{Merge to vectors of numbers in such a way that the output won't
contain numbers of the second vector that are within a interval
[i - window, i + window] for each number i in the first vector.
As an example, mergeRTs(c(1, 5), c(2, 3)) will result in c(1, 3, 5).}
\usage{
mergeRTs(rts1, rts2, window = 1)
}
\description{
Merge to vectors of numbers in such a way that the output won't
contain numbers of the second vector that are within a interval
[i - window, i + window] for each number i in the first vector.
As an example, mergeRTs(c(1, 5), c(2, 3)) will result in c(1, 3, 5).
}

