% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlsGaussianSum.R
\name{nlsGaussianSum}
\alias{nlsGaussianSum}
\title{Fit a nls model using multiple Gaussian functions.}
\usage{
nlsGaussianSum(y, x = seq_along(y), n = 1, spar = 0.5,
  with.plot = FALSE)
}
\arguments{
\item{y}{A numeric vector of intensity/signal values.}

\item{x}{An optional vector of x values.}

\item{n}{The number of components to fit.}

\item{spar}{A smoothness parameter that is passed to smooth.spline. Used to
presmooth the data.}

\item{with.plot}{If diagnostic plots should be made.}
}
\value{
The fitted nls model.
}
\description{
Fit a nls model using multiple Gaussian functions.
}

