% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlsGaussianSum.R
\name{nlsGaussianSumCV}
\alias{nlsGaussianSumCV}
\title{Choose the best number of gaussian components automatically using K-fold
cross validation. The best performing model is refitted and returned.}
\usage{
nlsGaussianSumCV(y, x = seq_along(y), K = 3, n.range = 1:3,
  quiet = TRUE, ...)
}
\arguments{
\item{y}{The intensity values on which to fit.}

\item{x}{The associated x values. Optional.}

\item{K}{The number of cross validation folds to perform. Larger numbers
yield better error estimates.}

\item{n.range}{A numeric vector of n values that should be evaluated.}
}
\value{
The fitted nls model that performed best.
}
\description{
Choose the best number of gaussian components automatically using K-fold
cross validation. The best performing model is refitted and returned.
}

