% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slidingWindowSubgroupDetection.R
\name{plot.swResult}
\alias{plot.swResult}
\title{Plot the result of the slidingWindow algorithm.}
\usage{
plot.swResult(sw.result, protein.traces.long, n.largest = NULL, log = FALSE)
}
\description{
Plot the result of the slidingWindow algorithm.
}

