% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manualComplexFeatureAnnotation.R
\name{plotROC}
\alias{plotROC}
\title{Plot a ROC curve.}
\usage{
plotROC(df, dynamic.axis = TRUE)
}
\arguments{
\item{df}{A dataframe as returned by makeROCWithSEC}

\item{dynamic.axis}{If the axis should be adjusted s.t. only visible values
are shown.}
}
\description{
Plot a ROC curve.
}

