% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benchmarkAnalysisPC.R
\name{plotSiteLocalizationRate}
\alias{plotSiteLocalizationRate}
\title{Plot the site localization rate for the output of `calculateDetectionStats`.}
\usage{
plotSiteLocalizationRate(stats, index)
}
\description{
Plot the site localization rate for the output of `calculateDetectionStats`.
}

