% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataVisualization.R
\name{plotTracesWithFeaturesCC}
\alias{plotTracesWithFeaturesCC}
\title{Plot traces and features for CC wf.}
\usage{
plotTracesWithFeaturesCC(detected.features.cc,
  complex.assoc = corum.complex.protein.assoc, traces.cc, pdf.file.loc,
  nplots = NULL)
}
\arguments{
\item{detected.features.cc}{A datatable of the output from openms.}

\item{complex.assoc}{The complex <-> protein associations.}

\item{traces.cc}{The data.file that was input into the cprophet CC wf.}

\item{pdf.file.loc}{The location where to store the pdf.}
}
\description{
Plot traces and features for CC wf.
}

