% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlsGaussianSum.R
\name{predict.nlsGaussianSum}
\alias{predict.nlsGaussianSum}
\title{Predict using a nlsGaussianSum model.}
\usage{
\method{predict}{nlsGaussianSum}(fit, x)
}
\description{
Predict using a nlsGaussianSum model.
}

