% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataInput.R
\name{produceComplexAnnotatedProteinTraces}
\alias{produceComplexAnnotatedProteinTraces}
\title{Combine a list of complex <-> protein associations and observations of
peptide intensities into a list of protein intensities that are annotated
with the complex they MIGHT belong to.}
\usage{
produceComplexAnnotatedProteinTraces(filename.corum.complex.assoc,
  filename.peptide.traces.long)
}
\arguments{
\item{filename.corum.complex.assoc}{The file to the association TSV file.
This table must have the columns: 'complex_id', 'protein_id'.}

\item{filename.peptide.traces.long}{The file to the peptide intensity
       This table must have the columns: 'peptide_id', 'sec', 'peptide_intensity',
       and 'protein_id'. 'sec' corresponds to a numeric time value and 'protein_id'
       to the protein that was inferred for this peptide.}
}
\value{
A data.table of protein intensity observations.
}
\description{
Combine a list of complex <-> protein associations and observations of
peptide intensities into a list of protein intensities that are annotated
with the complex they MIGHT belong to.
}

