% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataConversion.R
\name{produceProteinTraces}
\alias{produceProteinTraces}
\title{Sum up a list of peptide intensities into a list of protein intensities.}
\usage{
produceProteinTraces(peptide.traces)
}
\arguments{
\item{peptide.traces}{A long format peptide trace data.table.}
}
\value{
A data.table of protein intensity observations with.
        The DT will have the columns: 'protein_id', 'sec', 'intensity'.
}
\description{
Sum up a list of peptide intensities into a list of protein intensities.
}
\examples{
peptraces <- widePepTracesToLong(e4.peptide.traces.wide.filtered)
prottraces <- produceProteinTraces(peptraces)
}

