% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein.traces.2peps}
\alias{protein.traces.2peps}
\title{Summed up protein traces using peptides from experiment 4 with the settings:
\itemize{
 \item m-score less than 1 percent
 \item no requant
 \item no decoy
 \item only 2 most intense peptides summed up
}}
\format{\preformatted{Classes ‘data.table’ and 'data.frame':	5 obs. of  83 variables:
 $ protein_id: chr  "O00116" "O00148" "O00161" "O00161" ...
 $ complex_id: chr  "635" "5615" "1169" "730" ...
 $ 3         : int  0 0 0 0 0
 $ 4         : int  0 0 0 0 0
 $ 5         : int  0 2208 0 0 0
 $ 6         : int  4100 0 6250 6250 6250
 $ 7         : int  10802 0 0 0 0
 $ 8         : int  16534 0 6097 6097 6097
 $ 9         : int  17828 0 7163 7163 7163
 $ 10        : int  20596 0 6876 6876 6876
 $ 11        : int  26654 2652 12627 12627 12627
 $ 12        : int  32719 0 8098 8098 8098
 $ 13        : int  37633 0 12804 12804 12804
 $ 14        : int  32269 0 6699 6699 6699
 $ 15        : int  35472 0 6704 6704 6704
 $ 16        : int  30352 0 5253 5253 5253
 $ 17        : int  26295 0 7418 7418 7418
 $ 18        : int  33964 663 4757 4757 4757
 $ 19        : int  45592 0 9410 9410 9410
 $ 20        : int  35075 0 3677 3677 3677
 $ 21        : int  39766 0 4785 4785 4785
 $ 22        : int  34361 0 3973 3973 3973
 $ 23        : int  33279 0 5117 5117 5117
 $ 24        : int  29352 0 4257 4257 4257
 $ 25        : int  21192 0 8147 8147 8147
 $ 26        : int  26789 2738 8413 8413 8413
 $ 27        : int  18915 0 6129 6129 6129
 $ 28        : int  16446 0 5323 5323 5323
 $ 29        : int  16851 0 5062 5062 5062
 $ 30        : int  8813 0 3511 3511 3511
 $ 31        : int  12744 0 3526 3526 3526
 $ 32        : int  7633 0 0 0 0
 $ 33        : int  7785 0 1428 1428 1428
 $ 34        : int  8152 0 0 0 0
 $ 35        : int  6115 2151 1870 1870 1870
 $ 36        : int  0 0 0 0 0
 $ 37        : int  10468 0 0 0 0
 $ 38        : int  3665 0 0 0 0
 $ 39        : int  7439 0 0 0 0
 $ 40        : int  8119 0 0 0 0
 $ 41        : int  9231 0 0 0 0
 $ 42        : int  9864 0 0 0 0
 $ 43        : int  5163 0 1987 1987 1987
 $ 44        : int  12138 2271 0 0 0
 $ 45        : int  11964 2678 2376 2376 2376
 $ 46        : int  12139 2939 0 0 0
 $ 47        : int  14544 2573 0 0 0
 $ 48        : int  15683 5706 0 0 0
 $ 49        : int  0 6056 2650 2650 2650
 $ 50        : int  6118 6172 4052 4052 4052
 $ 51        : int  0 4865 6097 6097 6097
 $ 52        : int  5619 5760 7598 7598 7598
 $ 53        : int  9548 24533 7224 7224 7224
 $ 54        : int  7083 103044 5137 5137 5137
 $ 55        : int  0 172525 3737 3737 3737
 $ 56        : int  7067 200389 0 0 0
 $ 57        : int  5505 111812 0 0 0
 $ 58        : int  6121 66417 0 0 0
 $ 59        : int  7577 43222 4776 4776 4776
 $ 60        : int  8383 31569 0 0 0
 $ 61        : int  7333 16309 0 0 0
 $ 62        : int  11885 17392 0 0 0
 $ 63        : int  15907 26802 0 0 0
 $ 64        : int  15642 23629 0 0 0
 $ 65        : int  12913 16509 1042 1042 1042
 $ 66        : int  11131 14883 1024 1024 1024
 $ 67        : int  10578 27643 743 743 743
 $ 68        : int  10755 20028 961 961 961
 $ 69        : int  4859 11725 742 742 742
 $ 70        : int  8990 15564 0 0 0
 $ 71        : int  5671 14895 0 0 0
 $ 72        : int  6491 18214 0 0 0
 $ 73        : int  5193 16654 0 0 0
 $ 74        : int  2033 5639 0 0 0
 $ 75        : int  2672 10146 0 0 0
 $ 76        : int  7325 20967 0 0 0
 $ 77        : int  1268 4409 0 0 0
 $ 78        : int  4284 11414 0 0 0
 $ 79        : int  5158 10565 0 0 0
 $ 80        : int  3620 9960 0 0 0
 $ 81        : int  3242 7654 0 0 0
 $ 82        : int  3077 6542 0 0 0
 $ 83        : int  2211 6541 0 0 0
 - attr(*, ".internal.selfref")=<externalptr> 
}}
\usage{
protein.traces.2peps
}
\description{
Summed up protein traces using peptides from experiment 4 with the settings:
\itemize{
 \item m-score less than 1 percent
 \item no requant
 \item no decoy
 \item only 2 most intense peptides summed up
}
}
\keyword{datasets}

