% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manualComplexFeatureAnnotation.R
\name{readManualAnnotationFile}
\alias{readManualAnnotationFile}
\title{Read a TSV file where each row corresponds to an annotated complex.
The table must have the columns: 'complex_id', 'apexes_fully_observed',
'apexes_partially_observed'. The numbers within the apex columns have to be
comma-separated and should not be surrounded by whitespace.}
\usage{
readManualAnnotationFile(fname)
}
\arguments{
\item{fname}{The filename of the TSV file.}
}
\value{
A data.table.
}
\description{
Read a TSV file where each row corresponds to an annotated complex.
The table must have the columns: 'complex_id', 'apexes_fully_observed',
'apexes_partially_observed'. The numbers within the apex columns have to be
comma-separated and should not be surrounded by whitespace.
}

