% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{removeStretchesFilter}
\alias{removeStretchesFilter}
\title{Remove consecutive stretches from a matrix of peptide traces.}
\usage{
removeStretchesFilter(mat, cutoff)
}
\arguments{
\item{x}{A single integer.}
}
\description{
Remove consecutive stretches from a matrix of peptide traces.
}

