% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/peptideFilters.R
\name{requireAtLeastXPeptides}
\alias{requireAtLeastXPeptides}
\title{Filter a list of peptide intensities s.t. only those are kept that belong to
peptides whose protein has at least `min.peptides` peptides.}
\usage{
requireAtLeastXPeptides(peps, min.peptides)
}
\arguments{
\item{peps}{A long format data.table of peptide intensities.
The table should have the following columns:
\itemize {
  \item protein_id (character)
  \item peptide_id (character)
  \item sec (integer)
  \item intensity (numeric)
}}
}
\description{
Filter a list of peptide intensities s.t. only those are kept that belong to
peptides whose protein has at least `min.peptides` peptides.
}

