% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/peptideFilters.R
\name{retainXMostIntensePeptides}
\alias{retainXMostIntensePeptides}
\title{Filter a list of peptide intensities s.t. only those peptides are kept that
belong to the `max.peptides` peptides for each protein that have the highest
total intensity.}
\usage{
retainXMostIntensePeptides(peps, max.peptides)
}
\arguments{
\item{peps}{A long format data.table of peptide intensities.
The table should have the following columns:
\itemize {
  \item protein_id (character)
  \item peptide_id (character)
  \item sec (integer)
  \item intensity (numeric)
}}
}
\description{
Filter a list of peptide intensities s.t. only those peptides are kept that
belong to the `max.peptides` peptides for each protein that have the highest
total intensity.
}

