% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slidingWindowCorrelation.R
\name{slidingWindowCorrelation}
\alias{slidingWindowCorrelation}
\title{Compute the correlation between rows in a matrix using a sliding window.}
\usage{
slidingWindowCorrelation(trace.mat, window.size, score = "pearson")
}
\arguments{
\item{trace.mat}{A numeric matrix where rows correspond to the different
traces.}

\item{window.size}{Size of the window. Numeric.}
}
\value{
A vector of length `ncol(trace.mat) where each element corresponds to
a mean correlation within the window at that position.
}
\description{
Compute the correlation between rows in a matrix using a sliding window.
}

