% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataInput.R
\name{widePepTracesToLong}
\alias{widePepTracesToLong}
\title{Convert a wide format peptide trace data.table to long format.
This functionc an be used to convert the included data.table
'e4.peptide.traces' to long format.}
\usage{
widePepTracesToLong(dt)
}
\arguments{
\item{dt}{The wide format data.table.}
}
\value{
The same data.table in long format.
}
\description{
Convert a wide format peptide trace data.table to long format.
This functionc an be used to convert the included data.table
'e4.peptide.traces' to long format.
}

