% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataConversion.R
\name{wideProtTracesToLong}
\alias{wideProtTracesToLong}
\title{Convert a wide format protein trace data.table to long format.}
\usage{
wideProtTracesToLong(dt)
}
\description{
Convert a wide format protein trace data.table to long format.
}

