% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corum.complex.protein.assoc}
\alias{corum.complex.protein.assoc}
\title{A data.table of associations between corum complexes and their
protein subunits. Complexes that have multiple possible subunit
compositions are split into separate complexes that are numbered
with `-1`, `-2`, etc. suffixes.}
\format{\preformatted{Classes ‘data.table’ and 'data.frame':	9701 obs. of  3 variables:
 $ complex_id  : chr  "1" "1" "10" "10" ...
 $ complex_name: chr  "BCL6-HDAC4 complex" "BCL6-HDAC4 complex" "13S condensin complex" "13S condensin complex" ...
 $ protein_id  : chr  "P41182" "P56524" "Q15021" "Q9BPX3" ...
 - attr(*, ".internal.selfref")=<externalptr> 
}}
\usage{
corum.complex.protein.assoc
}
\description{
A data.table of associations between corum complexes and their
protein subunits. Complexes that have multiple possible subunit
compositions are split into separate complexes that are numbered
with `-1`, `-2`, etc. suffixes.
}
\keyword{datasets}

