% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manual.annotations.partial.imputed.win14.corr0.7}
\alias{manual.annotations.partial.imputed.win14.corr0.7}
\title{Manual annotations of partial peak groups that whose completeness
was estimated by hierchical clustering of the trace correlation profiles.}
\format{\preformatted{Classes ‘data.table’ and 'data.frame':	1519 obs. of  5 variables:
 $ complex_id                    : chr  "5615" "5615" "59" "59" ...
 $ rt                            : num  7 52 18 35 20 19 35 55 20 61 ...
 $ n_proteins_in_complete_complex: int  19 19 5 5 5 13 13 36 36 15 ...
 $ n_proteins_in_complex         : int  6 6 3 3 3 7 7 14 14 5 ...
 $ completeness                  : chr  "partial" "partial" "partial" "partial" ...
 - attr(*, ".internal.selfref")=<externalptr> 
 - attr(*, "index")= atomic  
  ..- attr(*, "__complex_id")= int  784 1068 1069 1070 1071 1072 1051 1052 1053 1004 ...
}}
\usage{
manual.annotations.partial.imputed.win14.corr0.7
}
\description{
Manual annotations of partial peak groups that whose completeness
was estimated by hierchical clustering of the trace correlation profiles.
}
\keyword{datasets}

