% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manual.annotations.partial}
\alias{manual.annotations.partial}
\title{The merged manual CORUM annotations with those features that were partially
observed i.e. some of the proteins that should belong to the CORUM complex
did not take part in the peak group.}
\format{\preformatted{Classes ‘data.table’ and 'data.frame':	1519 obs. of  2 variables:
 $ complex_id: Factor w/ 587 levels "5615","59","652",..: 1 1 2 2 2 3 3 4 4 5 ...
 $ rt        : num  7 52 18 35 20 19 35 55 20 61 ...
 - attr(*, ".internal.selfref")=<externalptr> 
}}
\usage{
manual.annotations.partial
}
\description{
The merged manual CORUM annotations with those features that were partially
observed i.e. some of the proteins that should belong to the CORUM complex
did not take part in the peak group.
}
\keyword{datasets}

