% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manual.annotations.raw.mheusel}
\alias{manual.annotations.raw.mheusel}
\title{The raw manual complex annotations by mheusel
These annotations are based on the traces
summed up protein traces from
e4.peptide.traces.wide.filtered.
A already processed and merged  version exists in the data frame
\itemize{
 \item manual.annotations.partial
 \item manual.annotations.full
 \item manual.annotations.full.complete
}}
\format{\preformatted{Classes ‘data.table’ and 'data.frame':	1719 obs. of  5 variables:
 $ complex_id                    : chr  "635" "5615" "1169" "730" ...
 $ apexes (fully_observed)       : chr  "18,20" "" "12,21" "11" ...
 $ apexes (partially_observed)   : chr  "" "7,52" "" "" ...
 $ n_proteins_in_complex         : int  2 19 4 2 2 1 1 2 1 1 ...
 $ n_proteins_in_complete_complex: int  2 23 4 3 3 3 2 2 2 2 ...
 - attr(*, ".internal.selfref")=<externalptr> 
}}
\usage{
manual.annotations.raw.mheusel
}
\description{
The raw manual complex annotations by mheusel
These annotations are based on the traces
summed up protein traces from
e4.peptide.traces.wide.filtered.
A already processed and merged  version exists in the data frame
\itemize{
 \item manual.annotations.partial
 \item manual.annotations.full
 \item manual.annotations.full.complete
}
}
\keyword{datasets}

