% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manual.annotations.raw.rhafen}
\alias{manual.annotations.raw.rhafen}
\title{The raw manual complex annotations by rhafen.
These annotations are based on the traces
summed up protein traces from
e4.peptide.traces.wide.filtered.
A already processed and merged  version exists in the data frame
\itemize {
 \item manual.annotations.partial
 \item manual.annotations.full
 \item manual.annotations.full.complete
}}
\format{\preformatted{ chr "~/Dev/MAScripts/manual_complex_annotation/COMPLEXES_4_osw_output_mscore_lt_1percent_no_requant_no_decoy_FILTERED_ANNOTATED.tsv"
}}
\usage{
manual.annotations.raw.rhafen
}
\description{
The raw manual complex annotations by rhafen.
These annotations are based on the traces
summed up protein traces from
e4.peptide.traces.wide.filtered.
A already processed and merged  version exists in the data frame
\itemize {
 \item manual.annotations.partial
 \item manual.annotations.full
 \item manual.annotations.full.complete
}
}
\keyword{datasets}

