% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataInput.R
\name{produceComplexAnnotatedProteinTraces}
\alias{produceComplexAnnotatedProteinTraces}
\title{Combine a list of complex <-> protein associations and observations of
peptide intensities into a list of protein intensities that are annotated
with the complex they MIGHT belong to.}
\usage{
produceComplexAnnotatedProteinTraces(peptide.traces, corum.protein.assoc)
}
\arguments{
\item{peptide.traces}{A wide format peptide trace data.table.}

\item{corum.complex.assoc}{A DT of protein <-> corum associations.
This table must have the columns: 'complex_id', 'complex_name', protein_id'.}
}
\value{
A data.table of protein intensity observations with.
        The DT will have the columns: 'protein_id', 'sec', 'intensity',
        'complex_id', 'complex_name'.
}
\description{
Combine a list of complex <-> protein associations and observations of
peptide intensities into a list of protein intensities that are annotated
with the complex they MIGHT belong to.
}
\examples{
produceComplexAnnotatedProteinTraces(e4.peptide.traces.wide.filtered,
                                     corum.complex.protein.assoc)
}

